requirements.txt
linkedin_api
requests
pandas
numpy
pathlib
typing
datetime
logging
dataclasses
python-dotenv